Pre-final update
Added barebones of real-time data, added minor features, and tried making UI more appeasing to the eye. To-do: Add the final bar graph comparing particular city's AQI to other city's in the state.
