--- data/India/India.cpg
@@ -1,1 +1,0 @@
-UTF-8
